loading bundle resources + change objc app to use static library.
