+++ HyBidDemo/HyBidDemoObjc/HyBidStatic/OMSDK-1.2.20/OMSDK_Pubnativenet.framework/Headers
@@ -1,0 +1,1 @@
+Versions/Current/Headers
