+++ HyBidDemo/HyBidDemoObjc/HyBidStatic/OMSDK-1.2.20/OMSDK_Pubnativenet.framework/Modules/module.modulemap
@@ -1,0 +1,4 @@
+module OMSDK_Pubnativenet {
+    header "Headers/OMIDImports.h"
+    export *
+}
