--- HyBidDemo/HyBidDemoStatic/HyBidStatic/OMSDK-1.2.20/OMSDK_Pubnativenet.framework/Headers
@@ -1,1 +1,0 @@
-Versions/Current/Headers
