--- HyBidDemo/HyBidDemoStatic/HyBidStatic/OMSDK-1.2.20/OMSDK_Pubnativenet.framework/Versions/A/Headers/OMIDImports.h
@@ -1,9 +1,0 @@
-#import "OMIDSDK.h"
-#import "OMIDScriptInjector.h"
-#import "OMIDPartner.h"
-#import "OMIDVerificationScriptResource.h"
-#import "OMIDAdSessionContext.h"
-#import "OMIDAdSession.h"
-#import "OMIDAdEvents.h"
-#import "OMIDVASTProperties.h"
-#import "OMIDVideoEvents.h"
