+++ PubnativeLite/Frameworks/static/HyBidStatic.bundle/_CodeSignature/CodeResources
@@ -38,23 +38,23 @@
 		</data>
 		<key>PNLiteVASTPlayerFullScreenViewController.nib/objects-11.0+.nib</key>
 		<data>
-		docRD3hRWvmA3/A5NpsJXFm2JiQ=
+		EMgITEOOVD4tp94w8hAQEjkjMNE=
 		</data>
 		<key>PNLiteVASTPlayerFullScreenViewController.nib/runtime.nib</key>
 		<data>
-		PIsLpplDWhEFORwwjWEaSGoiGRA=
+		hhplP1g7ulFGeWq6BAyj4w8QlSI=
 		</data>
 		<key>PNLiteVASTPlayerInterstitialViewController.nib</key>
 		<data>
-		5fZJorby+u5bMCBvnbmr7dF8yrI=
+		FFOfkXqbmfeQB5+Tsv7L3nWgyW4=
 		</data>
 		<key>PNLiteVASTPlayerViewController.nib/objects-11.0+.nib</key>
 		<data>
-		llVjTJCxJtaFCkjqF6ZCMffi0ss=
+		VBu9mM6cacDaNukf8XWCvG7RD8M=
 		</data>
 		<key>PNLiteVASTPlayerViewController.nib/runtime.nib</key>
 		<data>
-		AUHQVQRetWS/XiTp6mCHSSmfdiE=
+		NpO1jT9i7OPACLSHwsy6VW13Hts=
 		</data>
 		<key>omsdk.js</key>
 		<data>
@@ -144,55 +144,55 @@
 		<dict>
 			<key>hash</key>
 			<data>
-			docRD3hRWvmA3/A5NpsJXFm2JiQ=
+			EMgITEOOVD4tp94w8hAQEjkjMNE=
 			</data>
 			<key>hash2</key>
 			<data>
-			0mEUP5ThB/4yS1TKkxKrQ2cEiFfGUDWbZsuDPS46yiA=
+			zickaf/xM2nrMpr+newZueyY0cPmiBbMSdqf7B+QvNw=
 			</data>
 		</dict>
 		<key>PNLiteVASTPlayerFullScreenViewController.nib/runtime.nib</key>
 		<dict>
 			<key>hash</key>
 			<data>
-			PIsLpplDWhEFORwwjWEaSGoiGRA=
+			hhplP1g7ulFGeWq6BAyj4w8QlSI=
 			</data>
 			<key>hash2</key>
 			<data>
-			vuc/StRJTlwk2NUk661nDXyauG7Zz70HMOgGINKRSoA=
+			9VaIH/ZD2tQAsbdwxPAQVaLbHEItaqL6d50g+agVDXY=
 			</data>
 		</dict>
 		<key>PNLiteVASTPlayerInterstitialViewController.nib</key>
 		<dict>
 			<key>hash</key>
 			<data>
-			5fZJorby+u5bMCBvnbmr7dF8yrI=
+			FFOfkXqbmfeQB5+Tsv7L3nWgyW4=
 			</data>
 			<key>hash2</key>
 			<data>
-			lRFVjybfjaJkOHdF6TfsRNDvipfmTBUSuHQTA8HALxU=
+			VcIrmHJzOH2kVdsYs4EzGYpLl4uvyktnFcB8Gj1aWkU=
 			</data>
 		</dict>
 		<key>PNLiteVASTPlayerViewController.nib/objects-11.0+.nib</key>
 		<dict>
 			<key>hash</key>
 			<data>
-			llVjTJCxJtaFCkjqF6ZCMffi0ss=
+			VBu9mM6cacDaNukf8XWCvG7RD8M=
 			</data>
 			<key>hash2</key>
 			<data>
-			+QfUI20CcbgjOoLX6wG64o2wD2Rpl20xyCnv7x+qa+8=
+			vB2nX7P/Jhk+6+BMvTl957m0i/4apGAiVNuJ5/egWVU=
 			</data>
 		</dict>
 		<key>PNLiteVASTPlayerViewController.nib/runtime.nib</key>
 		<dict>
 			<key>hash</key>
 			<data>
-			AUHQVQRetWS/XiTp6mCHSSmfdiE=
+			NpO1jT9i7OPACLSHwsy6VW13Hts=
 			</data>
 			<key>hash2</key>
 			<data>
-			hsnqnNKSqzlUJJSzt4U/w4YP/bWoquFTqvW8du1eCrU=
+			bqk5Yqn7ZEA771rA4DKT6A2WAedJf4BwSOZRCaZCm3k=
 			</data>
 		</dict>
 		<key>omsdk.js</key>
